specs for entriesOf inAnyOrder(.only)
