+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/MapContainsInAnyOrderEntriesOfAssertionsSpec.kt
@@ -1,32 +1,36 @@
-package ch.tutteli.atrium.api.infix.en_GB
+package ch.tutteli.atrium.api.fluent.en_GB
 
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.logic.creating.typeutils.MapLike
-import ch.tutteli.atrium.specs.*
-import ch.tutteli.atrium.specs.integration.mfun2
+import ch.tutteli.atrium.specs.withNullableSuffix
 import org.spekframework.spek2.Spek
 import kotlin.jvm.JvmName
-import ch.tutteli.atrium.api.infix.en_GB.MapContainsInAnyOrderEntriesOfAssertionsSpec.Companion as C
+import ch.tutteli.atrium.api.fluent.en_GB.MapContainsInAnyOrderEntriesOfAssertionsSpec.Companion as C
 
 class MapContainsInAnyOrderEntriesOfAssertionsSpec : Spek({
     include(BuilderSpec)
-    //TODO 0.15.0:
+    //TODO 0.15.0: enable once implemented
     //include(ShortcutSpec)
+
+    //TODO 0.15.0: enable once implemented
+    //include(MapLikeSpec)
 }) {
+
     object BuilderSpec : ch.tutteli.atrium.specs.integration.MapContainsInAnyOrderKeyValuePairsAssertionsSpec(
         containsKeyValuePair_s to C::containsKeyValuePairs,
         (containsKeyValuePair_s to C::containsKeyValuePairsNullable).withNullableSuffix(),
-        "[Atrium][Shortcut] "
+        "[Atrium][Builder] "
     )
 
-    object ShortcutSpec : ch.tutteli.atrium.specs.integration.MapContainsInAnyOrderKeyValuePairsAssertionsSpec(
-        mfun2<String, Int, Int>(C::contains),
-        mfun2<String?, Int?, Int?>(C::contains).withNullableSuffix(),
-        "[Atrium][Shortcut] "
-    )
+    //TODO 0.15.0: add MapLikeSpec which tests passing illegal values
+//    object MapLikeSpec : ch.tutteli.atrium.specs.integration.MapContainsInAnyOrderKeyValuePairsAssertionsSpec(
+//        mfun2<String, Int, Int>(Expect<Map<out String, Int>>::contains),
+//        mfun2<String?, Int?, Int?>(Expect<Map<out String?, Int?>>::contains).withNullableSuffix(),
+//        "[Atrium][Shortcut] "
+//    )
 
     companion object : MapContainsSpecBase() {
-        val containsKeyValuePair_s = "$contains $filler $inAnyOrder keyValue"
+        val containsKeyValuePair_s = "${contains}.${inAnyOrder}.$entriesOf"
 
         private fun containsKeyValuePairs(
             expect: Expect<Map<out String, Int>>,
@@ -38,7 +42,7 @@ class MapContainsInAnyOrderEntriesOfAssertionsSpec : Spek({
             } else {
                 mapOf(a, *aX)
             }
-            return expect contains o inAny order entriesOf mapLike
+            return expect.contains.inAnyOrder.entriesOf(mapLike)
         }
 
         private fun containsKeyValuePairsNullable(
@@ -46,21 +50,21 @@ class MapContainsInAnyOrderEntriesOfAssertionsSpec : Spek({
             a: Pair<String?, Int?>,
             aX: Array<out Pair<String?, Int?>>
         ): Expect<Map<out String?, Int?>> =
-            expect contains o inAny order entriesOf listOf(a, *aX)
+            expect.contains.inAnyOrder.entriesOf(listOf(a, *aX))
 
-        private fun contains(
+        private fun containsEntriesOf(
             expect: Expect<Map<out String, Int>>,
             a: Pair<String, Int>,
             aX: Array<out Pair<String, Int>>
         ): Expect<Map<out String, Int>> =
-            expect contains o inAny order entriesOf sequenceOf(a, *aX)
+            expect.contains.inAnyOrder.entriesOf(sequenceOf(a, *aX))
 
         @JvmName("containsNullable")
-        private fun contains(
+        private fun containsEntriesOf(
             expect: Expect<Map<out String?, Int?>>,
             a: Pair<String?, Int?>,
             aX: Array<out Pair<String?, Int?>>
         ): Expect<Map<out String?, Int?>> =
-            expect contains o inAny order entriesOf arrayOf(a, *aX)
+            expect.contains.inAnyOrder.entriesOf(arrayOf(a, *aX))
     }
 }
