+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/MapContainsInAnyOrderKeyValuePairsAssertionsSpec.kt
@@ -1,0 +1,47 @@
+package ch.tutteli.atrium.api.fluent.en_GB
+
+import ch.tutteli.atrium.creating.Expect
+import ch.tutteli.atrium.logic.utils.mapArguments
+import ch.tutteli.atrium.specs.*
+import ch.tutteli.atrium.specs.integration.mfun2
+import org.spekframework.spek2.Spek
+import kotlin.jvm.JvmName
+import ch.tutteli.atrium.api.fluent.en_GB.MapContainsInAnyOrderKeyValuePairsAssertionsSpec.Companion as C
+
+class MapContainsInAnyOrderKeyValuePairsAssertionsSpec : Spek({
+    include(BuilderSpec)
+    include(ShortcutSpec)
+}) {
+
+    object BuilderSpec : ch.tutteli.atrium.specs.integration.MapContainsInAnyOrderKeyValuePairsAssertionsSpec(
+        containsKeyValuePair_s to C::containsKeyValuePairs,
+        (containsKeyValuePair_s to C::containsKeyValuePairsNullable).withNullableSuffix(),
+        "[Atrium][Builder] "
+    )
+
+    object ShortcutSpec : ch.tutteli.atrium.specs.integration.MapContainsInAnyOrderKeyValuePairsAssertionsSpec(
+        mfun2<String, Int, Int>(Expect<Map<out String, Int>>::contains),
+        mfun2<String?, Int?, Int?>(Expect<Map<out String?, Int?>>::contains).withNullableSuffix(),
+        "[Atrium][Shortcut] "
+    )
+
+    companion object : MapContainsSpecBase() {
+        val containsKeyValuePair_s = "${contains}.${inAnyOrder}.$keyValuePair/$keyValuePairs"
+
+        private fun containsKeyValuePairs(
+            expect: Expect<Map<out String, Int>>,
+            a: Pair<String, Int>,
+            aX: Array<out Pair<String, Int>>
+        ): Expect<Map<out String, Int>> =
+            if (aX.isEmpty()) expect.contains.inAnyOrder.entry(a)
+            else expect.contains.inAnyOrder.entries(a, *aX)
+
+        private fun containsKeyValuePairsNullable(
+            expect: Expect<Map<out String?, Int?>>,
+            a: Pair<String?, Int?>,
+            aX: Array<out Pair<String?, Int?>>
+        ): Expect<Map<out String?, Int?>> =
+            if (aX.isEmpty()) expect.contains.inAnyOrder.entry(a)
+            else expect.contains.inAnyOrder.entries(a, *aX)
+    }
+}
