+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/MapContainsInAnyOrderOnlyEntriesOfAssertionsSpec.kt
@@ -1,0 +1,48 @@
+package ch.tutteli.atrium.api.fluent.en_GB
+
+import ch.tutteli.atrium.creating.Expect
+import ch.tutteli.atrium.logic.creating.typeutils.MapLike
+import ch.tutteli.atrium.specs.*
+import org.spekframework.spek2.Spek
+import ch.tutteli.atrium.api.fluent.en_GB.MapContainsInAnyOrderOnlyEntriesOfAssertionsSpec.Companion as C
+
+class MapContainsInAnyOrderOnlyEntriesOfAssertionsSpec : Spek({
+    include(BuilderSpec)
+    //TODO 0.15.0: enable once implemented
+    //include(ShortcutSpec)
+
+    //TODO 0.15.0: enable once implemented
+    //include(MapLikeSpec)
+}) {
+
+    object BuilderSpec : ch.tutteli.atrium.specs.integration.MapContainsInAnyOrderOnlyKeyValuePairsAssertionsSpec(
+        containsKeyValuePair_s to C::containsKeyValuePairs,
+        (containsKeyValuePair_s to C::containsKeyValuePairsNullable).withNullableSuffix(),
+        "◆ ", "✔ ", "✘ ", "❗❗ ", "⚬ ", "» ", "▶ ", "◾ ",
+        "[Atrium][Builder] "
+    )
+
+    companion object : MapContainsSpecBase() {
+        val containsKeyValuePair_s = "${contains}.${inAnyOrder}.${only}.$entriesOf"
+
+        private fun containsKeyValuePairs(
+            expect: Expect<Map<out String, Int>>,
+            a: Pair<String, Int>,
+            aX: Array<out Pair<String, Int>>
+        ): Expect<Map<out String, Int>> {
+            val mapLike: MapLike = if (aX.distinct().size != aX.size || aX.contains(a)) {
+                sequenceOf(a, *aX).asIterable()
+            } else {
+                mapOf(a, *aX)
+            }
+            return expect.contains.inAnyOrder.only.entriesOf(mapLike)
+        }
+
+        private fun containsKeyValuePairsNullable(
+            expect: Expect<Map<out String?, Int?>>,
+            a: Pair<String?, Int?>,
+            aX: Array<out Pair<String?, Int?>>
+        ): Expect<Map<out String?, Int?>> =
+            expect.contains.inAnyOrder.only.entriesOf(listOf(a, *aX))
+    }
+}
