+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/MapContainsInAnyOrderOnlyKeyValueAssertionsSpec.kt
@@ -3,20 +3,16 @@ package ch.tutteli.atrium.api.fluent.en_GB
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.logic.utils.mapArguments
 import ch.tutteli.atrium.specs.*
-import ch.tutteli.atrium.specs.integration.mfun2
 import org.spekframework.spek2.Spek
-import kotlin.jvm.JvmName
-import ch.tutteli.atrium.api.fluent.en_GB.MapContainsInAnyOrderOnlyAssertionsSpec.Companion as C
+import ch.tutteli.atrium.api.fluent.en_GB.MapContainsInAnyOrderOnlyKeyValueAssertionsSpec.Companion as C
 
-class MapContainsInAnyOrderOnlyAssertionsSpec : Spek({
+class MapContainsInAnyOrderOnlyKeyValueAssertionsSpec : Spek({
     include(BuilderSpec)
     //TODO #68 in 0.15.0
 //    include(ShortcutSpec)
 }) {
 
-    object BuilderSpec : ch.tutteli.atrium.specs.integration.MapContainsInAnyOrderOnlyAssertionsSpec(
-        containsKeyValuePair_s to C::containsKeyValuePairs,
-        (containsKeyValuePair_s to C::containsKeyValuePairsNullable).withNullableSuffix(),
+    object BuilderSpec : ch.tutteli.atrium.specs.integration.MapContainsInAnyOrderOnlyKeyValueAssertionsSpec(
         containsKeyValue_s to C::containsKeyValues,
         (containsKeyValue_s to C::containsKeyValuesNullable).withNullableSuffix(),
         "◆ ", "✔ ", "✘ ", "❗❗ ", "⚬ ", "» ", "▶ ", "◾ ",
@@ -34,24 +30,7 @@ class MapContainsInAnyOrderOnlyAssertionsSpec : Spek({
 //    )
 
     companion object : MapContainsSpecBase() {
-        val containsKeyValuePair_s = "${contains}.${inAnyOrder}.${only}.$keyValuePair/$keyValuePairs"
         val containsKeyValue_s = "${contains}.${inAnyOrder}.${only}.$keyValue/$keyValues"
-
-        private fun containsKeyValuePairs(
-            expect: Expect<Map<out String, Int>>,
-            a: Pair<String, Int>,
-            aX: Array<out Pair<String, Int>>
-        ): Expect<Map<out String, Int>> =
-            if (aX.isEmpty()) expect.contains.inAnyOrder.only.entry(a)
-            else expect.contains.inAnyOrder.only.entries(a, *aX)
-
-        private fun containsKeyValuePairsNullable(
-            expect: Expect<Map<out String?, Int?>>,
-            a: Pair<String?, Int?>,
-            aX: Array<out Pair<String?, Int?>>
-        ): Expect<Map<out String?, Int?>> =
-            if (aX.isEmpty()) expect.contains.inAnyOrder.only.entry(a)
-            else expect.contains.inAnyOrder.only.entries(a, *aX)
 
         private fun containsKeyValues(
             expect: Expect<Map<out String, Int>>,
