+++ apis/infix-en_GB/atrium-api-infix-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/infix/en_GB/mapLikeContainsInOrderOnlyCreators.kt
@@ -104,7 +104,7 @@ inline infix fun <K, reified V : Any, T : MapLike> EntryPointStep<K, out V?, T, 
  *
  * @since 0.15.0
  */
-inline infix fun <reified K, reified V, T : MapLike> EntryPointStep<K, V, T, InOrderOnlySearchBehaviour>.keyValuePairsOf(
+inline infix fun <reified K, reified V, T : MapLike> EntryPointStep<K, V, T, InOrderOnlySearchBehaviour>.entriesOf(
     expectedMapLike: MapLike
 ): Expect<T> = _logic.toVarArgPairs<K, V>(expectedMapLike).let { (first, rest) ->
     this the pairs(first, *rest)
