+++ misc/specs/atrium-specs-common/src/main/kotlin/ch/tutteli/atrium/specs/integration/MapContainsInAnyOrderKeyValueAssertionsSpec.kt
@@ -8,12 +8,6 @@ import ch.tutteli.atrium.translations.DescriptionMapAssertion
 import org.spekframework.spek2.style.specification.Suite
 import kotlin.reflect.KFunction3
 
-typealias MFun2<K, V, T> = Fun2<Map<out K, V>, Pair<K, T>, Array<out Pair<K, T>>>
-
-fun <K, V, T> mfun2(
-    f: KFunction3<Expect<Map<out K, V>>, Pair<K, T>, Array<out Pair<K, T>>, Expect<Map<out K, V>>>
-) = fun2(f)
-
 fun keyValue(key: String, assertionCreator: Expect<Int>.() -> Unit): Pair<String, Expect<Int>.() -> Unit> =
     key to assertionCreator
 
@@ -22,9 +16,7 @@ fun keyNullableValue(
     assertionCreator: (Expect<Int>.() -> Unit)?
 ): Pair<String?, (Expect<Int>.() -> Unit)?> = key to assertionCreator
 
-abstract class MapContainsInAnyOrderAssertionsSpec(
-    keyValuePairs: MFun2<String, Int, Int>,
-    keyValuePairsNullable: MFun2<String?, Int?, Int?>,
+abstract class MapContainsInAnyOrderKeyValueAssertionsSpec(
     keyWithValueAssertions: MFun2<String, Int, Expect<Int>.() -> Unit>,
     keyWithNullableValueAssertions: MFun2<String?, Int?, (Expect<Int>.() -> Unit)?>,
     describePrefix: String = "[Atrium] "
@@ -32,7 +24,6 @@ abstract class MapContainsInAnyOrderAssertionsSpec(
 
     include(object : SubjectLessSpec<Map<out String, Int>>(
         describePrefix,
-        keyValuePairs.forSubjectLess("key" to 1, arrayOf()),
         keyWithValueAssertions.forSubjectLess(
             keyValue("a") { toBe(1) },
             arrayOf(keyValue("a") { isLessThanOrEqual(2) })
@@ -41,7 +32,6 @@ abstract class MapContainsInAnyOrderAssertionsSpec(
 
     include(object : SubjectLessSpec<Map<out String?, Int?>>(
         "$describePrefix[nullable Key] ",
-        keyValuePairsNullable.forSubjectLess(null to 1, arrayOf("a" to null)),
         keyWithNullableValueAssertions.forSubjectLess(
             keyNullableValue(null) { toBe(1) },
             arrayOf(keyNullableValue("a", null))
@@ -71,97 +61,12 @@ abstract class MapContainsInAnyOrderAssertionsSpec(
     fun describeFun(vararg pairs: SpecPair<*>, body: Suite.() -> Unit) =
         describeFunTemplate(describePrefix, pairs.map { it.name }.toTypedArray(), body = body)
 
-    val fluent = expect(map)
-    val nullableFluent = expect(nullableMap)
-
-    fun entry(key: String): String = String.format(DescriptionMapAssertion.ENTRY_WITH_KEY.getDefault(), "\"$key\"")
-
-    fun entry(key: String, value: Any): String = entry(key) + ": " + value
-
-    describeFun(keyValuePairs, keyValuePairsNullable) {
-        val containsFunctions = uncheckedToNonNullable(keyValuePairs, keyValuePairsNullable)
-
-        context("map $map") {
-            containsFunctions.forEach { (name, containsFun) ->
-                listOf(
-                    listOf("a" to 1),
-                    listOf("b" to 2),
-                    listOf("a" to 1, "b" to 2),
-                    listOf("b" to 2, "a" to 1)
-                ).forEach {
-                    it("$name - $it does not throw") {
-                        fluent.containsFun(it.first(), it.drop(1).toTypedArray())
-                    }
-                }
-
-                it("$name - a to 1 and a to 1 does not throw (no unique match)") {
-                    fluent.containsFun("a" to 1, arrayOf("a" to 1))
-                }
-
-                it("$name - {a to 1, b to 3, c to 4} throws AssertionError, reports b and c") {
-                    expect {
-                        fluent.containsFun("a" to 1, arrayOf("b" to 3, "c" to 4))
-                    }.toThrow<AssertionError> {
-                        message {
-                            contains(
-                                entry("b", 2),
-                                "$toBeDescr: 3",
-                                entry("c", keyDoesNotExist),
-                                "$toBeDescr: 4"
-                            )
-                            containsNot(entry("a"))
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    describeFun(keyValuePairsNullable) {
-        val containsNullableFun = keyValuePairsNullable.lambda
-        context("map $nullableMap") {
-            listOf(
-                listOf("a" to null),
-                listOf(null to 1),
-                listOf("b" to 2),
-                listOf("a" to null, "b" to 2),
-                listOf(null to 1, "b" to 2),
-                listOf(null to 1, "a" to null),
-                listOf(null to 1, "a" to null, "b" to 2),
-                listOf("b" to 2, null to 1, "a" to null)
-            ).forEach {
-                it("$it does not throw") {
-                    nullableFluent.containsNullableFun(it.first(), it.drop(1).toTypedArray())
-                }
-            }
-
-            it("a to null and a to null does not throw (no unique match)") {
-                nullableFluent.containsNullableFun("a" to null, arrayOf("a" to null))
-            }
-
-            it("{a to null, null to 2, b to 3, c to 4} throws AssertionError, reports a, null, b and c") {
-                expect {
-                    nullableFluent.containsNullableFun("a" to null, arrayOf(null to 2, "b" to 3, "c" to 4))
-                }.toThrow<AssertionError> {
-                    message {
-                        contains(
-                            entry("b", 2),
-                            "$toBeDescr: 3",
-                            entry("c", keyDoesNotExist),
-                            "$toBeDescr: 4"
-                        )
-                        containsNot(entry("a"))
-                    }
-                }
-            }
-        }
-    }
-
     describeFun(keyWithValueAssertions, keyWithNullableValueAssertions) {
         val containsKeyWithValueAssertionsFunctions = uncheckedToNonNullable(
             keyWithValueAssertions,
             keyWithNullableValueAssertions
         )
+        val fluent = expect(map)
 
         context("map $map") {
             containsKeyWithValueAssertionsFunctions.forEach { (name, containsKeyWithValueAssertionsFun) ->
@@ -207,6 +112,8 @@ abstract class MapContainsInAnyOrderAssertionsSpec(
 
     describeFun(keyWithNullableValueAssertions) {
         val containsKeyWithNullableValueAssertionsFun = keyWithNullableValueAssertions.lambda
+        val nullableFluent = expect(nullableMap)
+
         context("map $nullableMap") {
             listOf(
                 "(a, null)" to
